requirements.txt
flask==0.12.2
SQLAlchemy==1.0.11
python-dotenv==0.6.4
requests==2.14.2
zappa>=0.17.6
flask_graphql==1.4.1
graphene_sqlalchemy==1.1.1